+++ src/tools/index.ts
@@ -3,6 +3,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 export { DelegateToAgentTool } from "./delegateToAgent";
+export { GeminiSearchTool } from "./geminiSearch";
 export { MiniMaxSearchTool } from "./minimaxSearch";
 export { registerAllTools } from "./registry";
 export { ZhipuSearchTool } from "./zhipuSearch";